feat(user-seen): add validation to prevent multiple entries deletion for the same date and view group

+++ libs/movies-prisma-lib/src/lib/delete-user-seen-entry.ts
@@ -5,6 +5,16 @@ export const deleteUserSeenEntry = async (args: {
   viewGroup: string;
   viewDate: Date;
 }) => {
+  const dateCount = await prisma.homewebbridge_userseen.count({
+    where: {
+      vdb_videoid: args.movieId,
+      asp_viewgroup: args.viewGroup,
+      viewdate: args.viewDate
+    }
+  });
+
+  if (dateCount > 1) throw new Error("Data inconsistency: Found two seen date for one view group at the same day.");
+
   return await prisma.homewebbridge_userseen.deleteMany({
     where: {
       vdb_videoid: args.movieId,
